Create README.md
Smartphones have become necessities in this modern age. People typically use smartphones for leisure activities; however, we have developed a mobile application that is practical as well as beneficial in that we aimed to measure and monitor vital signs in vulnerable people. The vast majority of cardiac diseases lead to mortality, according to the World Health Organization (WHO). Therefore, it is essential for people with cardiac diseases to keep track of their health problems and vital signs frequently.

The purposes of this study were to identify irregularities in heartbeats via the PanTompkins machine learning algorithm and develop a wearable device that operates through a mobile application.

The ECG data was collected from the wearable device and stored in Google Firebase for future usage. Then the data was sent to the mobile application for analysis. Next, the BPM was calculated and compared with the regular heart rate table. In real life, if the system were to detect an irregularity, it would send an emergency message to the individual’s family or a nearby doctor.